Time: 0 ms (100.00%) | Memory: 23.9 MB (45.52%) - LeetSync

+++ 3347-distribute-elements-into-two-arrays-i/distribute-elements-into-two-arrays-i.cpp
@@ -4,7 +4,7 @@ public:
         vector<int> arr1 = {nums[0]};
         vector<int> arr2 = {nums[1]};
 
-        for (size_t i = 2; i < nums.size(); ++i) {
+        for (int i = 2; i < nums.size(); ++i) {
             if (arr1.back() > arr2.back()) {
                 arr1.push_back(nums[i]);
             } else {
